chore: move from Utils folder to Constants folder
